requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==7.1.1
fastapi==0.115.6
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
groq==0.13.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.8.2
lancedb==0.17.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
mutagen==1.47.0
numpy==2.2.1
ollama==0.4.5
openai==1.58.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
phidata==2.7.5
platformdirs==4.3.6
primp==0.9.2
pyarrow==18.1.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pylance==0.20.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==12.6.0
sentry-sdk==1.45.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spotify_dl==8.9.0
spotipy==2.25.0
starlette==0.41.3
tantivy==0.22.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.34.0
webencodings==0.5.1
wheel==0.45.1
yfinance==0.2.51
youtube-transcript-api==0.6.3
yt-dlp==2024.12.23